+++ src/kanban/components/member/loadMembers.js
@@ -1,6 +1,6 @@
 import { api } from '../../../api/api.js';
 import localStorage from '../../../utils/localStorage.js';
-import { renderMember } from './renderMember.js';
+import { renderMember } from '../render/renderMember.js';
 
 export const loadMembers = async () => {
   const currentTeam = localStorage.get('currentTeam');
