+++ src/kanban/components/todo/filterTodo.js
@@ -1,0 +1,3 @@
+export const filterTodo = (target, currentTeam) => {
+  console.log(target);
+};
