+++ src/kanban/handler/handleClickTodoList.js
@@ -2,6 +2,7 @@ import { isContain } from '../../utils/validator.js';
 import { addMember } from '../components/member/addMember.js';
 import { deleteTodo } from '../components/todo/deleteTodo.js';
 import { deleteTodos } from '../components/todo/deleteTodos.js';
+import { filterTodo } from '../components/todo/filterTodo.js';
 import { toggleTodo } from '../components/todo/toggleTodo.js';
 
 export const handleClickTodoList = ({ target }, currentTeam) => {
@@ -19,5 +20,9 @@ export const handleClickTodoList = ({ target }, currentTeam) => {
   }
   if (isContain(target, 'clear-completed')) {
     deleteTodos(target, currentTeam);
+    return;
+  }
+  if (isContain(target, 'filter')) {
+    filterTodo(target, currentTeam);
   }
 };
